normalize some SIGNALs and SLOTs

+++ plugins/signalmonitor/signalhistorydelegate.cpp
@@ -69,9 +69,9 @@ void SignalHistoryDelegate::paint(QPainter *painter, const QStyleOptionViewItem 
     const QAbstractItemModel * const model = index.model();
     const QVector<qint64> &events
         = model->data(index, SignalHistoryModel::EventsRole).value<QVector<qint64> >();
-    const qint64 t0 = qMax(0LL, model->data(index,
-                                            SignalHistoryModel::StartTimeRole).value<qint64>()
-                           - startTime);
+    const qint64 t0
+        = qMax(static_cast<qint64>(0),
+               model->data(index, SignalHistoryModel::StartTimeRole).value<qint64>() - startTime);
     qint64 t1 = model->data(index, SignalHistoryModel::EndTimeRole).value<qint64>();
     if (t1 < 0) // still alive
         t1 = m_totalInterval;
